requirements.txt
pymongo
python-dotenv
requests
pydantic
# python-dotenv
langchain
langgraph
langchain-core
dnspython